move processor to a util file
