+++ src/routes/(site)/pages/[page]/+page.server.ts
@@ -1,18 +1,6 @@
 import { error } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
-import remarkGfm from 'remark-gfm';
-import { unified } from 'unified';
-import remarkParse from 'remark-parse';
-import remarkRehype from 'remark-rehype';
-import rehypeRaw from 'rehype-raw';
-import rehypeStringify from 'rehype-stringify';
-
-export const processor = unified()
-	.use(remarkParse)
-	.use(remarkGfm)
-	.use(remarkRehype, { allowDangerousHtml: true })
-	.use(rehypeRaw) // allow html within markdown files
-	.use(rehypeStringify);
+import { processor } from '$/utilities/markdown';
 
 export const load: PageServerLoad = async ({ params }) => {
 	const contentFiles = import.meta.glob('../*.md', { as: 'raw', eager: true });
